address review: isolate mamba paths, move HeteroTPTransferConfig to utils, fix _mamba_phys_ratio, refactor
